+++ src/main/java/io/github/itzispyder/util/MathUtil.java
@@ -5,15 +5,15 @@ import io.github.itzispyder.math.Vector;
 
 public class MathUtil {
 
-    public static double lerp(double a, double b, double delta) {
+    public static float lerp(float a, float b, float delta) {
         return a + (b - a) * delta;
     }
 
-    public static double lerp(double a, double b) {
+    public static float lerp(float a, float b) {
         return lerp(a, b, Main.tickDelta());
     }
 
-    public static Vector lerp(Vector a, Vector b, double delta) {
+    public static Vector lerp(Vector a, Vector b, float delta) {
         return a.add(b.sub(a).mul(delta));
     }
 
@@ -21,7 +21,7 @@ public class MathUtil {
         return lerp(a, b, Main.tickDelta());
     }
 
-    public static double clamp(double val, double min, double max) {
+    public static float clamp(float val, float min, float max) {
         return Math.max(min, Math.min(val, max));
     }
 }
